move Clock and timestamp generation/parsing into CompletionTracker
