+++ cf-ops-automation-broker-core/src/test/java/com/orange/oss/cloudfoundry/broker/opsautomation/ondemandbroker/terraform/TerraformCompletionTrackerTest.java
@@ -6,6 +6,9 @@ import org.springframework.cloud.servicebroker.model.GetLastServiceOperationResp
 import org.springframework.cloud.servicebroker.model.OperationState;
 
 import java.io.File;
+import java.time.Clock;
+import java.time.Instant;
+import java.time.ZoneId;
 
 import static org.fest.assertions.Assertions.assertThat;
 
@@ -25,7 +28,7 @@ public class TerraformCompletionTrackerTest {
         File tfStateFile = new File(path);
         assertThat(tfStateFile).exists();
 
-        TerraformCompletionTracker tracker = new TerraformCompletionTracker(tfStateFile);
+        TerraformCompletionTracker tracker = new TerraformCompletionTracker(tfStateFile, Clock.systemUTC());
 
         //when asked status
         GetLastServiceOperationResponse moduleExecStatus = tracker.getModuleExecStatus("4567");
@@ -44,9 +47,46 @@ public class TerraformCompletionTrackerTest {
         assert_expected_status(successOutput, successOutput, OperationState.SUCCEEDED);
     }
 
+    @Test
+    public void provides_current_date_for_last_operation() {
+        //given
+        Clock clock = Clock.fixed(Instant.ofEpochMilli(1510680248007L), ZoneId.of("Europe/Paris"));
+        TerraformCompletionTracker tracker = new TerraformCompletionTracker(Mockito.mock(File.class),clock);
+
+        //when
+        String currentDateAsOperation = tracker.getCurrentDate();
+        //then
+        assertThat(currentDateAsOperation).isEqualTo("2017-11-14T17:24:08.007Z");
+
+    }
+
+    @Test
+    public void get_elapsed_time_since_last_operation() {
+        //given
+        Clock clock = Clock.fixed(Instant.ofEpochMilli(1510680248007L+120*1000L), ZoneId.of("Europe/Paris"));
+        TerraformCompletionTracker tracker = new TerraformCompletionTracker(Mockito.mock(File.class),clock);
+
+        //when
+        long elapsedTimeSecsSinceLastOperation = tracker.getElapsedTimeSecsSinceLastOperation("2017-11-14T17:24:08.007Z");
+
+        //then
+        assertThat(elapsedTimeSecsSinceLastOperation).isEqualTo(120L);
+    }
+
+    @Test
+    public void given_we_understand_date_format_and_parsing() {
+        Clock clock = Clock.fixed(Instant.ofEpochMilli(1510680248007L), ZoneId.of("Europe/Paris"));
+
+        Instant now = Instant.now(clock);
+        assertThat(now.toString()).isEqualTo("2017-11-14T17:24:08.007Z");
+
+        assertThat(Instant.parse(now.toString())).isEqualTo(now);
+    }
+
+
     private void assert_expected_status(TerraformState.Output started, TerraformState.Output completed, OperationState expected) {
         File tfStateFile = Mockito.mock(File.class);
-        TerraformCompletionTracker tracker = new TerraformCompletionTracker(tfStateFile);
+        TerraformCompletionTracker tracker = new TerraformCompletionTracker(tfStateFile, Clock.systemUTC());
 
         GetLastServiceOperationResponse response = tracker.mapOutputToStatus(started, completed);
